feat: runtime paths enum
